Working on BBB implementation, logic functions well

+++ hal/beagleboneblack/src/bbb/hal.h
@@ -23,7 +23,7 @@ void BBB_shutdown();
  * DIO
 \***********************************************************************/
 
-hal_handle_t BBB_initializeDIOPort(uint8_t port, uint8_t dir);
+hal_handle_t BBB_initializeDIOPort(int16_t port, uint8_t dir);
 void BBB_freeDIOPort(hal_handle_t portHandle);
 
 void BBB_setDIO(hal_handle_t portHandle, uint8_t high);
@@ -35,7 +35,7 @@ uint8_t BBB_getDIO(hal_handle_t portHandle);
  * ANALOG
 \***********************************************************************/
 
-hal_handle_t BBB_initializeAnalogInput(uint8_t port);
+hal_handle_t BBB_initializeAnalogInput(int16_t port);
 void BBB_freeAnalogInput(hal_handle_t portHandle);
 
 uint32_t BBB_getAnalogValue(hal_handle_t portHandle);
@@ -45,7 +45,7 @@ float BBB_getAnalogVoltage(hal_handle_t portHandle);
  * PWM
 \***********************************************************************/
 
-hal_handle_t BBB_initializePWMPort(uint8_t port);
+hal_handle_t BBB_initializePWMPort(int16_t port);
 void BBB_freePWMPort(hal_handle_t portHandle);
 
 uint8_t BBB_getPWMValue(hal_handle_t portHandle);
